+++ libs/@livekit/components-react/src/components/controls/MediaDeviceSelect.tsx
@@ -67,6 +67,7 @@ export const MediaDeviceSelect: (
   ref
 ) {
   const room = useMaybeRoomContext();
+  const previousActiveDeviceId = React.useRef<string>('default');
   const handleError = React.useCallback(
     (e: Error) => {
       if (room) {
@@ -98,9 +99,10 @@ export const MediaDeviceSelect: (
   }, [onDeviceListChange, devices]);
 
   React.useEffect(() => {
-    if (activeDeviceId && activeDeviceId !== '') {
+    if (activeDeviceId !== previousActiveDeviceId.current) {
       onActiveDeviceChange?.(activeDeviceId);
     }
+    previousActiveDeviceId.current = activeDeviceId;
   }, [activeDeviceId]);
 
   const handleActiveDeviceChange = async (deviceId: string) => {