+++ libs/@livekit/components-react/src/hooks/useConnectionQualityIndicator.ts
@@ -1,6 +1,5 @@
 import { setupConnectionQualityIndicator } from '@livekit/components-core';
 import type { Participant } from 'livekit-client';
-import { ConnectionQuality } from 'livekit-client';
 import * as React from 'react';
 import { useEnsureParticipant } from '../context';
 import { useObservableState } from './internal';
@@ -32,7 +31,7 @@ export function useConnectionQualityIndicator(
 
   const quality = useObservableState(
     connectionQualityObserver,
-    ConnectionQuality.Unknown
+    p.connectionQuality
   );
 
   return { className, quality };